refactor AIAdatper out of ai telemetry appender
