+++ src/vs/workbench/parts/telemetry/test/node/appInsightsTelemetryAppender.test.ts
@@ -5,33 +5,35 @@
 'use strict';
 
 import * as assert from 'assert';
+import {IAIAdapter} from 'vs/base/node/aiAdapter';
 import { NodeAppInsightsTelemetryAppender } from 'vs/workbench/parts/telemetry/node/nodeAppInsightsTelemetryAppender';
 
 interface IAppInsightsEvent {
 	eventName: string;
-	properties?: {string?: string;};
-	measurements?: {string?: number;}
+	data: any;
 }
 
-class AppInsightsMock {
+class AIAdapterMock implements IAIAdapter {
 
 	public events: IAppInsightsEvent[]=[];
 	public IsTrackingPageView: boolean = false;
 	public exceptions: any[] =[];
 
-	public trackEvent(eventName: string, properties?: {string?: string;}, measurements?: {string?: number;}): void {
-		this.events.push({
-			eventName: eventName,
-			properties: properties,
-			measurements: measurements
-		});
-	}
-	public trackPageView(): void {
-		this.IsTrackingPageView = true;
+	constructor(private prefix: string, private eventPrefix: string, client?: any) {
 	}
 
-	public trackException(exception: any): void {
+	public log(eventName: string, data?: any): void {
+		this.events.push({
+			eventName: this.prefix+'/'+eventName,
+			data: data
+		});
+	}
+
+	public logException(exception: any): void {
 		this.exceptions.push(exception);
+	}
+
+	public dispose(): void {
 	}
 }
 
@@ -52,11 +54,11 @@ class ContextServiceMock {
 }
 
 suite('Telemetry - AppInsightsTelemetryAppender', () => {
-	var appInsightsMock: AppInsightsMock;
+	var appInsightsMock: AIAdapterMock;
 	var appender: NodeAppInsightsTelemetryAppender;
 
 	setup(() => {
-		appInsightsMock = new AppInsightsMock();
+		appInsightsMock = new AIAdapterMock(NodeAppInsightsTelemetryAppender.EVENT_NAME_PREFIX, NodeAppInsightsTelemetryAppender.EVENT_NAME_PREFIX);
 		appender = new NodeAppInsightsTelemetryAppender(null,<any> new ContextServiceMock('123'), appInsightsMock);
 	});
 
@@ -68,7 +70,7 @@ suite('Telemetry - AppInsightsTelemetryAppender', () => {
 		appender.log('testEvent');
 
 		assert.equal(appInsightsMock.events.length, 1);
-		assert.equal(appInsightsMock.events[0].eventName, NodeAppInsightsTelemetryAppender.EVENT_NAME_PREFIX+'testEvent');
+		assert.equal(appInsightsMock.events[0].eventName, NodeAppInsightsTelemetryAppender.EVENT_NAME_PREFIX+'/testEvent');
 	});
 
 	test('Track UnhandledError as exception and events', () => {
@@ -77,79 +79,25 @@ suite('Telemetry - AppInsightsTelemetryAppender', () => {
 		appender.log('UnhandledError', sampleError);
 
 		assert.equal(appInsightsMock.events.length, 1);
-		assert.equal(appInsightsMock.events[0].eventName, NodeAppInsightsTelemetryAppender.EVENT_NAME_PREFIX+'UnhandledError');
+		assert.equal(appInsightsMock.events[0].eventName, NodeAppInsightsTelemetryAppender.EVENT_NAME_PREFIX+'/UnhandledError');
 
 		assert.equal(appInsightsMock.exceptions.length, 1);
 	});
 
-	test('property limits', () => {
-		var reallyLongPropertyName = 'abcdefghijklmnopqrstuvwxyz';
-		for (var i =0; i <6; i++) {
-			reallyLongPropertyName +='abcdefghijklmnopqrstuvwxyz';
-		}
-		assert(reallyLongPropertyName.length > 150);
-
-		var reallyLongPropertyValue = 'abcdefghijklmnopqrstuvwxyz012345678901234567890123';
-		for (var i =0; i <21; i++) {
-			reallyLongPropertyValue +='abcdefghijklmnopqrstuvwxyz012345678901234567890123';
-		}
-		assert(reallyLongPropertyValue.length > 1024);
-
-		var data = {};
-		data[reallyLongPropertyName] = '1234';
-		data['reallyLongPropertyValue'] = reallyLongPropertyValue;
-		appender.log('testEvent', data);
-
-		assert.equal(appInsightsMock.events.length, 1);
-
-		for (var prop in appInsightsMock.events[0].properties){
-			assert(prop.length < 150);
-			assert(appInsightsMock.events[0].properties[prop].length <1024);
-		}
-	});
-
-	test('Different data types', () => {
-		var date = new Date();
-		appender.log('testEvent', { favoriteDate: date, likeRed: false, likeBlue: true, favoriteNumber:1,  favoriteColor: 'blue', favoriteCars: ['bmw', 'audi', 'ford']});
-
-		assert.equal(appInsightsMock.events.length, 1);
-		assert.equal(appInsightsMock.events[0].eventName, NodeAppInsightsTelemetryAppender.EVENT_NAME_PREFIX+'testEvent');
-		assert.equal(appInsightsMock.events[0].properties['favoriteColor'], 'blue');
-		assert.equal(appInsightsMock.events[0].measurements['likeRed'], 0);
-		assert.equal(appInsightsMock.events[0].measurements['likeBlue'], 1);
-		assert.equal(appInsightsMock.events[0].properties['favoriteDate'], date.toISOString());
-		assert.equal(appInsightsMock.events[0].properties['favoriteCars'], JSON.stringify(['bmw', 'audi', 'ford']));
-		assert.equal(appInsightsMock.events[0].measurements['favoriteNumber'], 1);
-	});
-
-	test('Nested data', () => {
+	test('Event with data', () => {
 		appender.log('testEvent', {
-			window : {
-				title: 'some title',
-				measurements: {
-					width: 100,
-					height: 200
-				}
-			},
-			nestedObj: {
-				nestedObj2: {
-					nestedObj3: {
-						testProperty: 'test',
-					}
-				},
-				testMeasurement:1
-			}
+			title: 'some title',
+			width: 100,
+			height: 200
 		});
 
 		assert.equal(appInsightsMock.events.length, 1);
-		assert.equal(appInsightsMock.events[0].eventName, NodeAppInsightsTelemetryAppender.EVENT_NAME_PREFIX+'testEvent');
+		assert.equal(appInsightsMock.events[0].eventName, NodeAppInsightsTelemetryAppender.EVENT_NAME_PREFIX+'/testEvent');
 
-		assert.equal(appInsightsMock.events[0].properties['window.title'], 'some title');
-		assert.equal(appInsightsMock.events[0].measurements['window.measurements.width'], 100);
-		assert.equal(appInsightsMock.events[0].measurements['window.measurements.height'], 200);
+		assert.equal(appInsightsMock.events[0].data['title'], 'some title');
+		assert.equal(appInsightsMock.events[0].data['width'], 100);
+		assert.equal(appInsightsMock.events[0].data['height'], 200);
 
-		assert.equal(appInsightsMock.events[0].properties['nestedObj.nestedObj2.nestedObj3'], JSON.stringify({"testProperty":"test"}));
-		assert.equal(appInsightsMock.events[0].measurements['nestedObj.testMeasurement'],1);
 	});
 
 	test('Test asimov', () => {
@@ -158,9 +106,9 @@ suite('Telemetry - AppInsightsTelemetryAppender', () => {
 		appender.log('testEvent');
 
 		assert.equal(appInsightsMock.events.length, 2);
-		assert.equal(appInsightsMock.events[0].eventName, NodeAppInsightsTelemetryAppender.EVENT_NAME_PREFIX+'testEvent');
+		assert.equal(appInsightsMock.events[0].eventName, NodeAppInsightsTelemetryAppender.EVENT_NAME_PREFIX+'/testEvent');
 
 		// test vortex
-		assert.equal(appInsightsMock.events[1].eventName, NodeAppInsightsTelemetryAppender.EVENT_NAME_PREFIX+'testEvent');
+		assert.equal(appInsightsMock.events[1].eventName, NodeAppInsightsTelemetryAppender.EVENT_NAME_PREFIX+'/testEvent');
 	});
 });
